feat: display no results found text dynamically
