+++ components/features/search/Search.tsx
@@ -1,21 +1,9 @@
-import { Dispatch, SetStateAction, useEffect, useState } from "react";
+import { useSearchContext } from "../../../context/SearchContext";
 
 import { Form, Input } from "./Search.styled";
 
-const Search: React.FC<{
-  setSearchResults: Dispatch<SetStateAction<any[]>>;
-}> = ({ setSearchResults }) => {
-  const [searchTerm, setSearchTerm] = useState<string>("");
-
-  async function getSearchResults() {
-    if (searchTerm === "") {
-      setSearchResults([]);
-    } else {
-      const res = await fetch(`/api/search?q=${searchTerm}`);
-      const { results } = await res.json();
-      setSearchResults(results);
-    }
-  }
+const Search: React.FC = () => {
+  const { searchTerm, setSearchTerm } = useSearchContext();
 
   function handleChange(e: any) {
     setSearchTerm(e.target.value.toLowerCase());
@@ -25,10 +13,6 @@ const Search: React.FC<{
     (e.keyCode === 13 || e.which === 13 || e.key === "Enter") &&
       e.preventDefault();
   }
-
-  useEffect(() => {
-    getSearchResults();
-  }, [searchTerm]);
 
   return (
     <Form>
